Began to add in the timer logic. The timing of the game will be contained within the timerlistener class. There will be an infinite loop running until there has been a hit or the missile reaches the ground.

+++ src/gameStuff/Missile.java
@@ -16,9 +16,9 @@ public class Missile {
 		return y;
 	}
 	
-	public void move(int xLoc, int yLoc) {
-		this.x = xLoc;
-		this.y = yLoc;
+	public void move(double xLoc, double yLoc) {
+		this.x = (int) xLoc;
+		this.y = (int) yLoc;
 	}
 
 }
